# internal/server/api/openapi.yaml
openapi: 3.1.0
info:
  title: crm-service
  version: 1.0.0
  description: CRM service API for managing customers.
servers:
  - url: http://localhost:8888
paths:
  /v1/customers:
    get:
      summary: List all customers
      operationId: listCustomers
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '400':
          description: Bad request
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'
        '400':
          description: Bad request
  /v1/customers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get customer by ID
      operationId: getCustomer
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Bad request
        '404':
          description: Customer not found
    patch:
      summary: Update customer by ID
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCustomerResponse'
        '400':
          description: Bad request
        '404':
          description: Customer not found
    delete:
      summary: Delete customer by ID
      operationId: deleteCustomer
      responses:
        '204':
          description: Customer deleted
        '400':
          description: Bad request
        '404':
          description: Customer not found

components:
  schemas:
    CreateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [firstName, lastName, email]
    CreateCustomerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
    UpdateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        contacted:
          type: boolean
    UpdateCustomerResponse:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [firstName, lastName, email]
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        contacted:
          type: boolean
      required: [firstName, lastName, email]